requirements.txt
requests>=2.31
pandas>=2.1
pyarrow>=14.0
google-cloud-storage>=2.13
google-cloud-bigquery>=3.13
python-dotenv>=1.0
python-dateutil>=2.8
